Avoid mutation of the plate_number public attribute
This change decouples PacBioEntity ID generation from the public
Pydantic API of the class. This avoids surprising mutation of the
plate_number attribute during validation.

A small number of tests checked for the mutation behaviour, so these
are also changed.

Remove redundant @classmethod decorators because the Pydantic
@field_validator already returns a class method.

Add exclusion directories to .flake8 to avoid false positives when
run locally.