Fixed a bug in generating the product ID.
An existing bug in constructing a JSON representation for
a Pydantic model when the plate_number attribute is 2 or
larger resulted in incorrect IDs being generated. This came
to light while testing https://github.com/wtsi-npg/npg_langqc
against release 5.0.0 of this package.

The fix includes a return to using Pydantic models' built-in
model_dump_json method in order to mitigate the risk that
comes manual generation of JSON strings.